requirements.txt
pandas~=1.1.3
scikit-learn~=0.23.2
joblib~=0.17.0
numpy~=1.19.2
python-dateutil~=2.8.1
matplotlib~=3.3.2
scipy~=1.5.2
nltk~=3.5
underthesea~=1.3.0
regex~=2020.11.13
pyvi~=0.1
fastapi~=0.63.0
starlette~=0.13.6
redis~=3.5.3
mongoengine~=0.22.1
pydantic~=1.7.3
uvicorn~=0.13.2
dynaconf~=3.1.2
python-dotenv~=0.15.0
celery~=5.0.5
ujson~=4.0.1